Refactored Wishlists and Behavior to Ports & Adapters, in order to hide the setters which are helpful for testing purposes, but should probably not be accessible to Santa in prod.
