+++ exercise/ts/day12/src/gifts/ForAssessingChildrensBehavior.ts
@@ -1,0 +1,7 @@
+import { Name } from './Name';
+import { Behavior } from './Behavior';
+
+export interface ForAssessingChildrensBehavior {
+    findBehaviorByName(childName: Name): Behavior;
+}
+
