+++ exercise/ts/day12/src/gifts/ForGettingWishlists.ts
@@ -1,0 +1,8 @@
+// src/gifts/ForGettingWishlists.ts
+import { Name } from './Name';
+import { Wishlist } from './Wishlist';
+
+export interface ForGettingWishlists {
+    findWishlistByChildName(childName: Name): Wishlist;
+}
+
