+++ exercise/ts/day12/tests/SantaBuilder.ts
@@ -1,24 +1,23 @@
+// tests/SantaBuilder.ts
 import { Santa } from '../src/gifts/Santa';
-import { BehaviorRepository } from '../src/gifts/BehaviorRepository';
-import { WishlistRepository } from '../src/gifts/WishlistRepository';
+import { BehaviorAssessmentForTest } from './BehaviorAssessmentForTest';
+import { WishlistRepositoryForTest } from './WishlistRepositoryForTest';
 import { Name } from '../src/gifts/Name';
 import { Behavior } from '../src/gifts/Behavior';
 import { Wishlist } from '../src/gifts/Wishlist';
 
 export class SantaBuilder {
-    private readonly behaviorRepository: BehaviorRepository;
-    private readonly wishlistRepository: WishlistRepository;
+    private readonly behaviorRepository: BehaviorAssessmentForTest;
+    private readonly wishlistRepository: WishlistRepositoryForTest;
 
     constructor() {
-        this.behaviorRepository = new BehaviorRepository();
-        this.wishlistRepository = new WishlistRepository();
+        this.behaviorRepository = new BehaviorAssessmentForTest();
+        this.wishlistRepository = new WishlistRepositoryForTest();
     }
 
     withChild(name: Name, behavior: Behavior, wishlist: Wishlist): SantaBuilder {
-        if (behavior)
-            this.behaviorRepository.setBehavior(name, behavior);
-        if (wishlist)
-            this.wishlistRepository.setWishlist(name, wishlist);
+        if (behavior) this.behaviorRepository.setBehavior(name, behavior);
+        if (wishlist) this.wishlistRepository.setWishlist(name, wishlist);
         return this;
     }
